Refactor 1

+++ OrderService/OrderService/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public Product(string productType, string productName, int price)
+        public Product(string productType, string productName, decimal price)
         {
             ProductType = productType;
             ProductName = productName;
@@ -11,12 +11,12 @@
 
         public string ProductType { get; }
         public string ProductName { get; }
-        public int Price { get; }
+        public decimal Price { get; }
 
         public class Prices
         {
-            public const int OneThousand = 1000;
-            public const int TwoThousand = 2000;
+            public const decimal OneThousand = 1000M;
+            public const decimal TwoThousand = 2000M;
             public const double TaxRate = .25d;
         }
     }
